requirements.txt
azure-functions
azure-identity
pdfplumber
sqlalchemy
pyodbc
python-tds
requests_toolbelt
requests
python-multipart